Call the transformer recursively, and pass it the Joi schema
My use case: I wanted to be able to attach JSON schema-specific
information to particular items in the Joi object, using Joi's `.meta()`
method. However, this wasn't possible, since `convert()` doesn't pass
the transformer to itself recursively. Furthermore, the transformer
needed access to the Joi schema object for the particular part of the
tree that it is transforming, so that it could inspect the `joi._meta`
property for the keys I'd added via `.meta()`.

This PR makes this possible by A) making sure the transform function is
passed around recursively, and B) passing the `joi` object to the
transform function.